# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FirmWizard API",
        "description": "AI-powered marketing platform for law firms",
        "contact": {
            "email": "support@firmwizard.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/ai-agents": {
            "get": {
                "tags": [
                    "AI Agents"
                ],
                "summary": "Get available AI agents",
                "operationId": "afb503bbf1ee5f5e7f555b0e6d089aa9",
                "responses": {
                    "200": {
                        "description": "List of available AI agents"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ai-agents/trigger/{agentType}": {
            "post": {
                "tags": [
                    "AI Agents"
                ],
                "summary": "Trigger an AI agent",
                "operationId": "6ef154775d6cab3e83a8a46aa9d25aa6",
                "parameters": [
                    {
                        "name": "agentType",
                        "in": "path",
                        "description": "Type of AI agent to trigger",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "blog_post",
                                "competitor_analysis",
                                "website_analysis",
                                "gbp_analysis",
                                "google_ads",
                                "lsa_ads",
                                "meta_ads",
                                "backlinks",
                                "analytics",
                                "firm_profile"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "input_data": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Agent triggered successfully"
                    },
                    "400": {
                        "description": "Invalid agent type"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ai-agents/jobs/{job}": {
            "get": {
                "tags": [
                    "AI Agents"
                ],
                "summary": "Get AI job results",
                "operationId": "eec596cc81637573504b78279d009928",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "description": "AI Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job details and results"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}